pyproject.toml
[tool.poetry]
name = "b2btest"
version = "0.1.0"
description = ""
authors = ["Titov Anton <webdev@titovanton.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
uvicorn = "^0.30.5"
gunicorn = "^22.0.0"
django-extensions = "^3.2.3"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
mypy = "^1.11.1"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
django-debug-toolbar = "^4.4.6"
django-stubs = "^5.0.4"
watchdog = "^4.0.1"
django-filter-stubs = "^0.1.3"
pytest-stub = "^1.1.0"
djangorestframework-stubs = "^3.15.0"
drf-spectacular = "^0.27.2"
drf-spectacular-sidecar = "^2024.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
